pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "law-rag"
version = "0.1.0"
description = "Vietnamese legal-document RAG with auto-ingest"
requires-python = ">=3.10"
dependencies = [
    "python-docx>=1.1.0",
    "watchdog>=4.0.0",
    "google-generativeai>=0.8.0",
    "FlagEmbedding>=1.2.10",
    "chromadb>=0.5.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.4.0",
    "rank_bm25>=0.2.2",
    "hdbscan>=0.8.33",
    "numpy>=1.26.0",
]

[tool.setuptools.packages.find]
where = ["src"]
